refactor(engine): extract spill/error.rs (SPL-2 #2324)
First step of the spill.rs decomposition plan. Moves the SpillError
enum and its Display/Error/From impls into a dedicated leaf module
under crates/engine/src/concurrent_delta/spill/error.rs.

spill.rs becomes spill/mod.rs and re-exports SpillError so the
existing public path `engine::concurrent_delta::spill::SpillError`
continues to resolve. No behavior or public API changes.
